Move Costs under the Clusters page folder
