+++ frontend/src/old-pages/Clusters/Costs/CostData.tsx
@@ -19,8 +19,8 @@ import {
 } from '@cloudscape-design/components'
 import {useCallback, useMemo} from 'react'
 import {useTranslation} from 'react-i18next'
-import EmptyState from '../../components/EmptyState'
-import {consoleDomain, useState} from '../../store'
+import EmptyState from '../../../components/EmptyState'
+import {consoleDomain, useState} from '../../../store'
 import {useCostMonitoringDataQuery} from './costs.queries'
 import {CostMonitoringData} from './costs.types'
 import {toShortDollarAmount} from './valueFormatter'
