+++ frontend/src/old-pages/Clusters/Costs/__tests__/valueFormatter.test.ts
